Conflicts:
	CHANGELOG.md
	RELEASE.md
	install/changes/2.3.3.txt

+++ framework/core/subsystems/expSession.php
@@ -1,7 +1,7 @@
 <?php
 ##################################################
 #
-# Copyright (c) 2004-2014 OIC Group, Inc.
+# Copyright (c) 2004-2015 OIC Group, Inc.
 #
 # This file is part of Exponent
 #
@@ -158,13 +158,15 @@ class expSession {
 	 */
 	public static function validate() {
 		global $db, $user;
+
 		//FJD - create a ticket for every session instead of just logged in users
 		if (empty($_SESSION[SYS_SESSION_KEY]['ticket'])) {
 			$ticket = self::createTicket();
 		} else {
 			$ticket = $db->selectObject('sessionticket',"ticket='".$_SESSION[SYS_SESSION_KEY]['ticket']."'");
 		}
-        if (empty($ticket)) $ticket = self::createTicket();
+        if (empty($ticket))
+            $ticket = self::createTicket();
 
 		//if we don't have a ticket here, that means the browser passed the cookie, the session is still
 		// active, but the DATABASE tickets table was cleared.
@@ -187,15 +189,15 @@ class expSession {
 			$user->id = 0;
 		}
 
-		if (!empty($ticket->refresh)) {
+		if (!empty($ticket->refresh)) {  // clear user session cache and reload permissions
 			if (isset($user)) expPermissions::load($user);
 			self::clearCurrentUserSessionCache();
 			$ticket->refresh = 0;
 		}
 
 		self::updateTicket($ticket, $user);
-		// Clean out old sessions from the sessionticket table.
-		$db->delete('sessionticket','last_active < ' . (time() - SESSION_TIMEOUT));
+        if (SESSION_TIMEOUT_ENABLE == true)
+            $db->delete('sessionticket','last_active < ' . (time() - SESSION_TIMEOUT));  // Clean out old sessions from the sessionticket table.
 
 		define('SITE_403_HTML', SITE_403_REAL_HTML);
 	}
@@ -210,6 +212,7 @@ class expSession {
 	 */
 	public static function updateTicket($ticket, $user){
 		global $db;
+
 		if (!empty($ticket->ticket)){
 			$ticket->uid = isset($user->id) ? $user->id : 0;
 			$ticket->last_active = time();
@@ -243,6 +246,7 @@ class expSession {
 		//	return $_SESSION[SYS_SESSION_KEY]['ticket'];
 		//} else return null;
 		global $db;
+
 		if (isset($_SESSION[SYS_SESSION_KEY]['ticket'])) {
 			if($db->selectObject('sessionticket',"ticket='".$_SESSION[SYS_SESSION_KEY]['ticket']."'") != null ) {
 				return $_SESSION[SYS_SESSION_KEY]['ticket'];
@@ -341,8 +345,9 @@ class expSession {
 		//at this point either.  This just updates all sessionticket records to refresh=1
 
 		global $db;
+
         $sessionticket = new stdClass();
-		$sessionticket->refresh = 1;
+		$sessionticket->refresh = 1;  // force user permissions & session cache reload
 		$db->updateObject($sessionticket, 'sessionticket', '1');
         self::clearCurrentUserSessionCache();
 
@@ -396,6 +401,7 @@ class expSession {
 	 */
 	public static function clearAllSessionData(){
 		global $db;
+
 		$db->delete('sessionticket',"1");
 		unset($_SESSION[SYS_SESSION_KEY]);
 	}
@@ -422,6 +428,7 @@ class expSession {
 		$_SESSION[SYS_SESSION_KEY]['ticket'] = $ticket->ticket;
 
 		global $db;
+
 		$db->insertObject($ticket,'sessionticket');
 		return $ticket;
 	}
@@ -430,6 +437,38 @@ class expSession {
 		//should always be an array, even if single index
 		$_SESSION[SYS_SESSION_KEY]['cache'][$module] = $val;
 	}
+
+    /** exdoc
+   	 * This call will force all active session to reload their
+   	 * permission data.  This is useful if permissions are assigned
+   	 * or revoked, and is required to see these changes.
+        *
+   	 * @node Subsystems:expPermissions
+   	 */
+   	public static function triggerRefresh() {
+   		global $db;
+
+   		$obj = new stdClass();
+   		$obj->refresh = 1;  // force user permissions & session cache reload
+   		$db->updateObject($obj,'sessionticket','true'); // force a global refresh
+   	}
+
+   	/** exdoc
+   	 * This call will force all active sessions for the given user to
+   	 * reload their permission data.  This is useful if permissions
+   	 * are assigned or revoked, and is required to see these changes.
+        *
+        * @param user/object $user
+        *
+   	 * @node Subsystems:expPermissions
+   	 */
+   	public static function triggerSingleRefresh($user) {  //FIXME not currently used
+   		global $db;
+
+   		$obj = new stdClass();
+   		$obj->refresh = 1;  // force user permissions & session cache reload
+   		$db->updateObject($obj,'sessionticket','uid='.$user->id); // force a global refresh
+   	}
 
 }
